requirements.txt
flask==3.0.3
pandas==1.3.5
joblib==0.17.0
python-whois
tldextract==5.1.2
pytz==2020.4
gunicorn
psycopg2-binary==2.9.5
SQLAlchemy==1.4.46
google-cloud-storage
validators
scikit-learn==1.0.2
numpy==1.21.6
scipy==1.7.3